Fix harmless compiler warnings coming from MSVC64 in test code.  No changes to the core.
FossilOrigin-Name: 274d2a22660c7b34b8bbd85f3c29cbafbcb1b4e7

+++ src/test_fs.c
@@ -99,7 +99,7 @@ static int fsConnect(
   }
   zTbl = argv[3];
 
-  nByte = sizeof(fs_vtab) + strlen(zTbl) + 1 + strlen(zDb) + 1;
+  nByte = sizeof(fs_vtab) + (int)strlen(zTbl) + 1 + (int)strlen(zDb) + 1;
   pVtab = (fs_vtab *)sqlite3MallocZero( nByte );
   if( !pVtab ) return SQLITE_NOMEM;
